Endpoint /fornecedores implementado

+++ src/main/java/com/gobra/sistemagestaodeobras/dto/LocalUsoResponseDTO.java
@@ -6,12 +6,18 @@ package com.gobra.sistemagestaodeobras.dto;
 import com.gobra.sistemagestaodeobras.model.LocalUso;
 
 public record LocalUsoResponseDTO(
-  long codigoLocalUsoObra,
+
+  Integer codigoLocalUsoObra,
+  
   String nomeLocalUsoObra
-) {
+
+  ) {
 
   public LocalUsoResponseDTO(LocalUso localUsoModel) {
     // O Lombok gerou os getters e setters dos atributos
-    this(localUsoModel.getCodigoLocalUsoObra(), localUsoModel.getNomeLocalUsoObra());
+    this(
+      localUsoModel.getCodigoLocalUsoObra(),
+      localUsoModel.getNomeLocalUsoObra()
+    );
   }
 }
